Ignore the content of the lab01/Images folder
All the files in the lab01/Images folder should be ignored by Git.

Changes in large binary files cause Git repositories to grow by the
size of the file in question every time the file is changed and the
change is committed. The growth directly affects the amount of data end
users have to retrieve when they need to clone the repository.