Plot two lines indicating the abscissa of extrema
It is useful to have two lines indicating the abscissa of extrema in
the two dimensional function (1st laboratory). In this way, it easier
to check whether or not they correspond to the values computed
analitically.

In this case, the lines y = -0.8165 and y = 0.8165 are plotted: these
values are the approximate roots of the equation 3x^2 - 2 = 0.