Add a script for the 1st part of the 2nd report
The script performs the third experiment of the gradient descent
methods section of the second report. This experiment focuses on
different functions which have gradiets pointing in the same direction:
different values of alpha must be used in order to minimize the
functions. Another way to solve the problem is to use the normalized
gradient: in this way you can use the same value of alpha for the
different functions.