Add a script for the Rosenbrock function section
This script performs some experiments regarding the Rosenbrock function
of the second report. It plots the function in order to give a general
idea of its features. Moreover, using different thresholds, it tries to
apply the backtracking gradient descent method in order to find the
global minimum of the function.