Add a script which finds the minimum quickly
The script is created by the Professor, and it uses a built-in function
of Python: line_search (part of scipy). It is an implementation of a
fast line search algorithm.

I need this script to compare its results with the results that I
computed using the backtracking method that I developed.