Add some features to line search (Rosenbrock)
I added some functionalities to the script that applies the line search
method (SciPy) to the Rosenbrock function (second report). It is now
able to draw the path made by the points found by the line search
method. Moreover, it saves all the results in some images, so that you
can see the main features of this method.