Fix backtracking approach used in Newton method
In the previous version of the script, I was trying to find a proper
value of alpha at each iteration using the backtracking approach of the
gradient descent method. Instead, I have to use the backtracking
approach applied to the Newton method, which is slightly different.

I've also fixed the part of the report related to this script, changing
the results that I obtained and how I should interpret them.