Study 1st lab function using Newton method
In the second report you can now find a comparison between the gradient
descent method and the Newton method. Moreover, I developed and
described a new method, which use the Newton approach whenever it's
possible. If it's not possible (because the Hessian matrix is not
positive definite) then the gradient descent method is used.