Apply Newton method to the Rosenbrock function
This script applies the Newton descent method (whenever it's possible,
that is whenever the Hessian is positive definite) to the Rosenbrock
function in the last part of the second report. It plots the points
computed by the algorithm with two different colors, so that you can
check which of the methods are used at each iteration.