Find a closed-form solution for the least squares
The script whose task was to perform a visual analysis has been renamed
so that it can contains all the code related to the least squares
problem of the third laboratory.

In the script I added a part that finds a closed-form solution to the
least squares problem using matrix algebra.